Changed how we're handling the insertion of a five line gap in poem 4. Instead of using the gap element alone, I added five lines, each containing a gap element.
Changed how we're handling line-omission. Instead of surrounding the
entire line with an app element, I'm putting the app element inside the
l element and making a rdg with gap. My reasoning is that the line is
present in some witnesses and sources, but it has been omitted in
others. For those, we should show that there is a gap where others have
something.